refactor useTheme and themeStore
