+++ next/src/layout/sidebar.tsx
@@ -15,7 +15,6 @@ import AppHead from "../components/AppHead";
 import LinkItem from "../components/sidebar/LinkItem";
 import AuthItem from "../components/sidebar/AuthItem";
 import { PAGE_LINKS, SOCIAL_LINKS } from "../components/sidebar/links";
-import { useThemeStore } from "../stores";
 import { useTheme } from "../hooks/useTheme";
 
 const SidebarLayout = (props: PropsWithChildren) => {
@@ -29,8 +28,7 @@ const SidebarLayout = (props: PropsWithChildren) => {
   });
   const userAgents = data ?? [];
 
-  const theme = useThemeStore.use.theme();
-  useTheme(theme);
+  useTheme();
 
   useEffect(() => {
     const handleResize = () => {
